requirements.txt
pycryptodome
pysmt
pytest
pytest-benchmark
pytest-profiling
pytest-timeout
pytest-xdist
solc-select
